+++ frontend/src/features/cardDetailedView/cardDetailedViewModal.tsx
